+++ src/extension.ts
@@ -69,7 +69,6 @@ export async function activate(context: vscode.ExtensionContext) {
 	);
 	ThisExtension.setStatusBar("Kernels initialized!");
 
-
 	// register DatabricksConnectionTreeProvider
 	let databricksConnectionTreeProvider = new DatabricksConnectionTreeProvider();
 	vscode.window.registerTreeDataProvider('databricksConnections', databricksConnectionTreeProvider);
@@ -81,7 +80,6 @@ export async function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand('databricksConnections.settings', () => databricksConnectionTreeProvider.openSettings());
 	}
 	vscode.commands.registerCommand('databricksConnectionItem.activate', (connection: DatabricksConnectionTreeItem) => connection.activate());
-
 
 	// register DatabricksWorkspaceTreeProvider
 	if (!ThisExtension.isInBrowser) {
@@ -102,11 +100,11 @@ export async function activate(context: vscode.ExtensionContext) {
 	// Workspace File System Provider
 	const workspaceProvider = new DatabricksWorkspaceProvider(context);
 	vscode.commands.registerCommand('databricksWorkspace.addToWorkspace', (showMessage: boolean) => {
-		FSHelper.addToWorkspace(vscode.Uri.parse('dbws:/'), "Databricks - Workspace", showMessage);
+		FSHelper.addToWorkspace(vscode.Uri.parse(ThisExtension.WORKSPACE_SCHEME +':/'), "Databricks - Workspace", showMessage);
 	});
 
 	if (ThisExtension.isInBrowser) {
-		// in a virtual workspace we always want to add the DBWS mount to the workspace
+		// in a virtual workspace we always want to add the WSFS/DBWS mount to the workspace
 		vscode.commands.executeCommand('databricksWorkspace.addToWorkspace', false);
 	}
 
@@ -126,6 +124,9 @@ export async function activate(context: vscode.ExtensionContext) {
 	vscode.commands.registerCommand('databricksClusterItem.useForSQL', (cluster: DatabricksCluster) => cluster.useForSQL());
 	vscode.commands.registerCommand('databricksClusterItem.createKernel', (cluster: DatabricksCluster) => cluster.createKernel());
 	vscode.commands.registerCommand('databricksClusterItem.restartKernel', (cluster: DatabricksCluster) => cluster.restartKernel());
+
+	// set cluster for Databricks Extension
+	vscode.commands.registerCommand('databricksClusterItem.attachCluster', (cluster: DatabricksCluster) => cluster.attachCluster());
 
 
 	// register DatabricksJobsTreeProvider
@@ -159,11 +160,11 @@ export async function activate(context: vscode.ExtensionContext) {
 	// DBFS File System Provider
 	const dbfsProvider = new DatabricksFileSystemProvider(context);
 	vscode.commands.registerCommand('databricksFS.addToWorkspace', (showMessage: boolean) => {
-		FSHelper.addToWorkspace(vscode.Uri.parse('dbfs:/'), "Databricks - DBFS", showMessage);
+		FSHelper.addToWorkspace(vscode.Uri.parse(ThisExtension.DBFS_SCHEME + ':/'), "Databricks - DBFS", showMessage);
 	});
 
 	if (ThisExtension.isInBrowser) {
-		// in a virtual workspace we always want to add the DBWS mount to the workspace
+		// in a virtual workspace we always want to add the DBFS mount to the workspace
 		vscode.commands.executeCommand('databricksFS.addToWorkspace', false);
 	}