+++ src/vscode/notebook/DatabricksKernel.ts
@@ -48,7 +48,7 @@ export class DatabricksKernel implements vscode.NotebookController {
 		this.notebookType = notebookType;
 		this._cluster = cluster;
 		this._language = language;
-		this.id = DatabricksKernel.getId(cluster.cluster_id, notebookType);
+		this.id = DatabricksKernel.getId(this.KernelID, notebookType);
 		this.label = DatabricksKernel.getLabel(cluster.cluster_name);
 
 		this._executionOrder = 0;
@@ -73,8 +73,8 @@ export class DatabricksKernel implements vscode.NotebookController {
 	}
 
 	async _onDidOpenNotebookDocument(notebook: vscode.NotebookDocument) {
-		// set this controller as recommended Kernel for notebooks opened via dbws:/ file system or from or local sync folder
-		if (notebook.uri.scheme == "dbws" || notebook.uri.toString().startsWith(ThisExtension.ActiveConnection.localSyncFolder.toString())) {
+		// set this controller as recommended Kernel for notebooks opened via dbws:/, wsfs:/ file system or from or local sync folder
+		if (notebook.uri.scheme == ThisExtension.WORKSPACE_SCHEME_LEGACY || notebook.uri.scheme == ThisExtension.WORKSPACE_SCHEME || notebook.uri.toString().startsWith(ThisExtension.ActiveConnection.localSyncFolder.toString())) {
 			this.Controller.updateNotebookAffinity(notebook, vscode.NotebookControllerAffinity.Preferred);
 		}
 
@@ -106,8 +106,8 @@ export class DatabricksKernel implements vscode.NotebookController {
 	}
 
 	// #region Cluster-properties
-	static getId(clusterId: string, kernelType: KernelType) {
-		return this.baseId + clusterId + "-" + kernelType;
+	static getId(kernelId: string, kernelType: KernelType) {
+		return kernelId + "-" + kernelType;
 	}
 
 	static getLabel(clusterName: string) {
@@ -116,6 +116,10 @@ export class DatabricksKernel implements vscode.NotebookController {
 
 	get Controller(): vscode.NotebookController {
 		return this._controller;
+	}
+
+	get KernelID(): string {
+		return this._cluster.kernel_id ?? this._cluster.cluster_id;
 	}
 
 	get ClusterID(): string {
@@ -387,9 +391,10 @@ export class DatabricksKernel implements vscode.NotebookController {
 					}
 					else { // absolute path provided
 						switch (cell.notebook.uri.scheme) {
-							case "dbws":
+							case ThisExtension.WORKSPACE_SCHEME_LEGACY: // legacy
+							case ThisExtension.WORKSPACE_SCHEME:
 								runUri = vscode.Uri.file(runFile);
-								runUri = runUri.with({ scheme: "dbws" })
+								runUri = runUri.with({ scheme: ThisExtension.WORKSPACE_SCHEME })
 								break;
 							case "file":
 								runUri = await FSHelper.joinPath(ThisExtension.ActiveConnection.localSyncFolder, ThisExtension.ActiveConnection.localSyncSubfolders.Workspace, runFile);
@@ -402,7 +407,8 @@ export class DatabricksKernel implements vscode.NotebookController {
 					ThisExtension.log("Executing %run for '" + runUri + "' ...");
 					try {
 						switch (runUri.scheme) {
-							case "dbws":
+							case ThisExtension.WORKSPACE_SCHEME_LEGACY: // legacy
+							case ThisExtension.WORKSPACE_SCHEME:
 								if (!await FSHelper.pathExists(runUri)) {
 									throw vscode.FileSystemError.FileNotFound(runUri);
 								}