+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+**v2.0.0:**
+- added integration with [official Databricks extensions](https://marketplace.visualstudio.com/items?itemName=databricks.databricks)
+  - new connection manager [Databricks Extensions](README.md/#setup-and-configuration-databricks-extension-connection-manager)
+  - derive cluster for [SQL Browser](README.md/#sql-browser)
+  - change cluster using [Cluster Manager](README.md/#cluster-manager)
+  - automatically create a [Notebook Kernel](README.md/#notebook-kernel) for the configured cluster
+  - added File System `wsfs:/` to replace `dbws:/` in the future (currently both are still supported)
+
 **v1.5.0:**
 - added support for [Widgets](README.md/#widgets) when running Notebooks